Move System Bridge services to dedicated file
